fix: make PR mergeable — split fields/primitives and fix Storybook import
