+++ ui/scripts/feature-cross-import-allowlist.mjs
@@ -630,19 +630,9 @@ export const allowlistedCrossFeatureBoundaryViolations = [
   },
   {
     relativeFilePath:
-      "src/features/current-selection/worker-selection/components/editable/worker-editable-configuration-section.tsx",
-    importSpecifiers: [
-      "../../../../current-factory-definition/public",
-    ],
-    reason:
-      "Legacy cross-feature internal import. Route reuse through the target feature public/ boundary or relocate shared behavior.",
-  },
-  {
-    relativeFilePath:
       "src/features/current-selection/worker-selection/components/editable/fields/worker-editable-configuration-shared-fields.tsx",
     importSpecifiers: [
       "../../../../../current-factory-definition/lib/worker-timeout-duration",
-      "../../../../../current-factory-definition/public",
     ],
     reason:
       "Legacy cross-feature internal import. Route reuse through the target feature public/ boundary or relocate shared behavior.",
