+++ ui/src/features/current-selection/worker-selection/components/editable/fields/worker-editable-configuration-shared-fields.tsx
@@ -5,12 +5,12 @@ import type {
   ReadyWorkerEditableConfigurationState,
   ReadyWorkerEditableValidationErrors,
   WorkerEditableConfigurationMessages,
-} from "./worker-editable-configuration-field-types";
+} from "./primitives/worker-editable-configuration-field-types";
 import {
   WorkerEditableConfigurationField,
   WorkerEditableConfigurationFieldHelp,
   WorkerEditableConfigurationServerChangedHint,
-} from "./worker-editable-configuration-field-primitives";
+} from "./primitives/worker-editable-configuration-field-primitives";
 
 export function WorkerEditableConfigurationSharedFields({
   messages,
